MouseTracker.js and Index files updated 05-04-21
Mouse tracking scripts moved to the MouseTracker.js file from the Index file.

X and Y coords as well as the Counter have been moved from the page itself to the console.

MouseTracker.js functioning, detects mouse movement (whether it is at coords 0,0 or undefined) as well as tracks whether the mouse is moving every 500 milliseconds.

So, index file is empty and MouseTracker.js is functioning.